[chore] #1 - 컴포넌트 분리
