+++ app/src/main/java/org/sopt/at/components/InputField.kt
@@ -1,0 +1,72 @@
+package org.sopt.at.components
+
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Call
+import androidx.compose.material.icons.filled.Notifications
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextField
+import androidx.compose.material3.TextFieldDefaults
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.input.PasswordVisualTransformation
+import androidx.compose.ui.text.input.VisualTransformation
+
+@Composable
+fun InputField(
+    value: String,
+    onValueChange: (String) -> Unit,
+    placeholder: String,
+    isPassword: Boolean = false,
+) {
+    var passwordVisible by remember { mutableStateOf(false) }
+
+    val visualTransformation = when {
+        isPassword && !passwordVisible -> PasswordVisualTransformation()
+        else -> VisualTransformation.None
+    }
+
+    TextField(
+        modifier = Modifier.fillMaxWidth(),
+        value = value,
+        onValueChange = onValueChange,
+        visualTransformation = visualTransformation,
+        placeholder = {
+            Text(text = placeholder)
+        },
+        colors = TextFieldDefaults.colors(
+            focusedTextColor = Color.LightGray,
+            unfocusedTextColor = Color.LightGray,
+            unfocusedContainerColor = Color(0xFF2C2C2C),
+            focusedContainerColor = Color(0xFF2C2C2C),
+            cursorColor = Color.LightGray,
+            unfocusedIndicatorColor = Color.Transparent,
+            focusedIndicatorColor = Color.Transparent,
+            disabledIndicatorColor = Color.Transparent
+        ),
+        trailingIcon = {
+            if (isPassword) {
+                val icon = if (passwordVisible) {
+                    Icons.Default.Notifications
+                } else {
+                    Icons.Default.Call
+                }
+
+                IconButton(onClick = { passwordVisible = !passwordVisible }) {
+                    Icon(
+                        imageVector = icon,
+                        contentDescription = if (passwordVisible) "비밀번호 숨기기" else "비밀번호 보기",
+                        tint = Color.White.copy(alpha = 0.6f)
+                    )
+                }
+            }
+        }
+    )
+}
